UPDATE:
after long time a small update.
Code to compile with the latest libraries.
Checked solution for beta=10, Bethe (D=1) U=5 against IPT.
Results are in good agreement for everything but the tail of the Self-energy.
Can't still figure out why. Probably the tails of both G0 and G are bad, just
like in CT-QMC.