v0.0.4 UPDATE - NEW VERSION:
This is a new corrected version. There was a huge error in the previous code
which went totally unnoticed in the single band case tested so far.
All the quantities have to be evaluated in the EigenBasis of the local
Hamiltonian, while previously FockBasis was used. The two coincided in the
single band case, but not in the multi-orbital one, causing the code
to produce wrong results.

This version is updated to work in the EigenBasis. Some work can be done
to optimize the calculation, but this is post-poned for the future optimization.

There is still a significant error in the evaluation of the density for the spin DW
channel in orbital 1. I still can not figure out why. This is temporary solved by
using Nspin=1 as a variable. Yet this patch does not work well for the double
occupancy, which comes out unbalanced on the two orbitals.

More tests to come, waiting for Angelo Valli.