v0.0.5 UPDATE - BUG FIX:
We solved the bug in the local observables calculation.
The solution is to create a local operator from its specific action
on the eigen-basis and then evaluate its expected value as a trace with
respect to R(beta).
We erroneously wrote the operator starting from the mat-mat product of
C and C^+ operators.